requirements.txt
google-generativeai
SpeechRecognition
streamlit
pyttsx3
